+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/dcql/DCQLJsonClaimsQuery.kt
@@ -66,7 +66,7 @@ data class DCQLJsonClaimsQuery(
             CredentialFormatEnum.JWT_VC,
         )
     ): KmmResult<DCQLClaimsQueryResult.JsonResult> = catching {
-        if (credentialQuery.format.coerceDeprecations() !in jsonBasedCredentialFormats) {
+        if (credentialQuery.format !in jsonBasedCredentialFormats) {
             throw IllegalArgumentException("Inconsistent credential format and claims query")
         }